refactor(pipeline/network): extract Scoring from NetworkDiscovery
Move endpoint-pool scorers, pickers, and discovery probes out of
NetworkDiscovery.ts into a dedicated Scoring/ sub-module:

  • Shape-aware tier picker (tierPick / discoverShapeAware) and its
    helpers (logShapeAwarePick, stampTierMeta, ITierPickOutcome,
    ShapeAwareTier, WINDOW_FROM_KEYS / WINDOW_TO_KEYS / hasWindowParams
    / safeParseWindowUrl).
  • Header probes — discoverByWellKnown, extractHeader,
    discoverHeaderValue, findByReferer, checkCorsHeader,
    findByCorsOrigin.
  • SPA URL discovery — CONFIG_URL_REGEX / INFRA_PREFIXES /
    isSpaCandidate, scanConfigBody, findByConfigBody,
    discoverSpaUrlFromTraffic.
  • API origin discovery — API_PATH_REGEX, extractApiFromBody,
    discoverApiFromConfig / Subdomain / Path,
    discoverApiOriginFromTraffic.
  • SPA-header carry-over — isBrowserStandard, extractSpaHeaders,
    spaHasAny.
  • findCommonServicesUrl — base-URL frequency picker.

NetworkDiscovery.ts imports the symbols it still needs
(discoverShapeAware / discoverHeaderValue / discoverByWellKnown /
discoverSpaUrlFromTraffic / discoverApiOriginFromTraffic /
extractSpaHeaders / spaHasAny / findCommonServicesUrl) from
Scoring/Scoring.js. Public surface unchanged — Scoring symbols are
internal to the Network module.

Phase 4 split: commit 4/9. NetworkDiscovery.ts: 1332 → 723 LoC.
